Add code, docker(-compose), nginx, certbot

+++ backend/models/Comment.js
@@ -34,8 +34,14 @@ Comment.handlers = {
         // data: 프론트 측에서 삭제 요청과 함께 온 데이터. data.id에 삭제 버튼을 클릭한 Comment id 담겨있음
         // name_by_auth: 요청을 보낸 사람의 이름
 
+        const comment = Comment.query.where({id: data.id})
+        if (name_by_auth !== 'wheel' && name_by_auth !== comment.name) {
+            return false;
+        }
 
-        const where = {/*TODO*/} //Hint: 5줄 이하, ORM(sequelize)에서 삭제 조건 어떻게 주는지 찾아보기
+        const where = {where: {
+            id: comment.id
+          }} //Hint: 5줄 이하, ORM(sequelize)에서 삭제 조건 어떻게 주는지 찾아보기
         await Comment.destroy({where})
             .catch(err => {
                 console.log(err);
